issue/55 Interim refactor
