+++ js/SearchableWord.js
@@ -1,0 +1,13 @@
+export default class SearchableWord {
+
+  constructor({
+    word,
+    score,
+    count
+  } = {}) {
+    this.word = word;
+    this.score = score;
+    this.count = count;
+  }
+
+}
