+++ js/searchDrawerItemView.js
@@ -8,31 +8,30 @@ export default class SearchDrawerItemView extends Backbone.View {
 
   events() {
     return {
-      'click .js-search-textbox-change': 'search',
-      'keyup .js-search-textbox-change': 'search'
+      'click .js-search-textbox-change': this.onSearch,
+      'keyup .js-search-textbox-change': this.onSearch
     };
   }
 
   initialize(options) {
     this.listenTo(Adapt, 'drawer:empty', this.remove);
     this.render();
-    this.search = _.debounce(this.search.bind(this), 1000);
-    if (options.query) {
-      this.$('.js-search-textbox-change').val(options.query);
-    }
+    this.search = _.debounce(this.onSearch.bind(this), 1000);
+    if (!options.query) return;
+    this.$('.js-search-textbox-change').val(options.query);
   }
 
   render() {
     const data = this.model.toJSON();
     const template = Handlebars.templates.searchBox;
-    $(this.el).html(template(data));
+    this.$el.html(template(data));
     return this;
   }
 
-  search(event) {
+  onSearch(event) {
     if (event && event.preventDefault) event.preventDefault();
-    const searchVal = this.$('.js-search-textbox-change').val();
-    Adapt.trigger('search:filterTerms', searchVal);
+    const query = this.$('.js-search-textbox-change').val();
+    Adapt.trigger('search:query', query);
   }
 
 }
