+++ js/searchResultsView.js
@@ -1,10 +1,6 @@
 import Adapt from 'core/js/adapt';
-import SearchAlgorithm from './search-algorithm';
-
-const replaceTagsRegEx = /<{1}[^>]+>/g;
-const replaceEscapedTagsRegEx = /&lt;[^&gt;]+&gt;/g;
-const replaceEndTagsRegEx = /<{1}\/{1}[^>]+>/g;
-const replaceEscapedEndTagsRegEx = /&lt;\/[^&gt;]+&gt;/g;
+import WORD_CHARACTERS from './WORD_CHARACTERS';
+import { stripTags, tag, escapeRegExp } from './utils';
 
 export default class SearchResultsView extends Backbone.View {
 
@@ -18,17 +14,14 @@ export default class SearchResultsView extends Backbone.View {
     };
   }
 
-  initialize(options) {
+  initialize({ searchObject = null } = {}) {
     this.listenTo(Adapt, {
       'drawer:empty': this.remove,
-      'search:termsFiltered': this.updateResults
+      'search:queried': this.updateResults
     });
-
     this.render();
-
-    if (options.searchObject) {
-      this.updateResults(options.searchObject);
-    }
+    if (!searchObject) return;
+    this.updateResults(searchObject);
   }
 
   render() {
@@ -44,138 +37,73 @@ export default class SearchResultsView extends Backbone.View {
   }
 
   formatResults(searchObject) {
-    const self = this;
     const resultsLimit = Math.min(5, searchObject.searchResults.length);
-
-    const formattedResults = _.map(_.first(searchObject.searchResults, resultsLimit), function(result) {
-      return self.formatResult(result);
-    });
-
-    searchObject.formattedResults = formattedResults;
-    return searchObject;
-  }
-
-  formatResult(result, query) {
-    const foundWords = _.keys(result.foundWords).join(' ');
-    let title = result.model.get('title');
-    let displayTitle = result.model.get('displayTitle');
-    let body = result.model.get('body');
     const previewWords = this.model.get('_previewWords');
     const previewCharacters = this.model.get('_previewCharacters');
-    const wordCharacters = window.search._regularExpressions.wordCharacters;
+    const formattedResults = searchObject.searchResults.slice(0, resultsLimit).map(result => {
+      const foundWords = _.keys(result.foundWords).join(' ');
+      const title = stripTags(result.model.get('title'));
+      const displayTitle = stripTags(result.model.get('displayTitle'));
+      const body = stripTags(result.model.get('body'));
+      const searchTitle = !title
+        ? $('<div>' + displayTitle + '</div>').text() || 'No title found'
+        : $('<div>' + title + '</div>').text();
+      let textPreview = '';
+      let finder;
+      // select preview text
+      if (result.foundPhrases.length > 0) {
+        let phrase = stripTags(result.foundPhrases[0].phrase);
+        let lowerPhrase = phrase.toLowerCase();
+        const lowerSearchTitle = searchTitle.toLowerCase();
 
-    // trim whitespace
-    title = title.replace(SearchAlgorithm._regularExpressions.trimReplaceWhitespace, '');
-    displayTitle = displayTitle.replace(SearchAlgorithm._regularExpressions.trimReplaceWhitespace, '');
-    body = body.replace(SearchAlgorithm._regularExpressions.trimReplaceWhitespace, '');
+        if (lowerPhrase === lowerSearchTitle && result.foundPhrases.length > 1) {
+          phrase = stripTags(result.foundPhrases[1].phrase);
+          lowerPhrase = phrase.toLowerCase();
+        }
 
-    // strip tags
-    title = this.stripTags(title);
-    displayTitle = this.stripTags(displayTitle);
-    body = this.stripTags(body);
+        if (lowerPhrase === lowerSearchTitle) {
+          // if the search phrase and title are the same
+          finder = new RegExp('(([^' + WORD_CHARACTERS + ']*[' + WORD_CHARACTERS + ']{1}){1,' + previewWords + '}|.{0,' + previewCharacters + '})', 'i');
+          if (body) {
+            textPreview = body.match(finder)[0] + '...';
+          }
+        } else {
+          const wordMap = Object.entries(result.foundWords).map(([word, count]) => ({ word, count }));
+          wordMap.sort((a, b) => a.count - b.count);
+          let wordIndex = 0;
+          let wordInPhraseStartPosition = lowerPhrase.indexOf(wordMap[wordIndex].word);
+          while (wordInPhraseStartPosition === -1) {
+            wordIndex++;
+            if (wordIndex === wordMap.length) throw new Error('search: cannot find word in phrase');
+            wordInPhraseStartPosition = lowerPhrase.indexOf(wordMap[wordIndex].word);
+          }
+          const regex = new RegExp('(([^' + WORD_CHARACTERS + ']*[' + WORD_CHARACTERS + ']{1}){1,' + previewWords + '}|.{0,' + previewCharacters + '})' + escapeRegExp(wordMap[wordIndex].word) + '(([' + WORD_CHARACTERS + ']{1}[^' + WORD_CHARACTERS + ']*){1,' + previewWords + '}|.{0,' + previewCharacters + '})', 'i');
+          const snippet = phrase.match(regex)[0];
+          const snippetIndexInPhrase = phrase.indexOf(snippet);
+          textPreview = `${(snippetIndexInPhrase !== 0) ? '...' : ''}${snippet}${(snippetIndexInPhrase + snippet.length !== phrase.length) ? '...' : ''}`;
+        }
 
-    let searchTitle = '';
-    let textPreview = '';
-
-    // select title
-    if (!title) {
-      searchTitle = $('<div>' + displayTitle + '</div>').text() || 'No title found';
-    } else {
-      searchTitle = $('<div>' + title + '</div>').text();
-    }
-
-    let finder;
-    // select preview text
-    if (result.foundPhrases.length > 0) {
-      let phrase = result.foundPhrases[0].phrase;
-      // strip tags
-      phrase = this.stripTags(phrase);
-
-      let lowerPhrase = phrase.toLowerCase();
-      const lowerSearchTitle = searchTitle.toLowerCase();
-
-      if (lowerPhrase === lowerSearchTitle && result.foundPhrases.length > 1) {
-        phrase = result.foundPhrases[1].phrase;
-        // strip tags
-        phrase = this.stripTags(phrase);
-        lowerPhrase = phrase.toLowerCase();
-      }
-
-      if (lowerPhrase === lowerSearchTitle) {
-        // if the search phrase and title are the same
-        finder = new RegExp('(([^' + wordCharacters + ']*[' + wordCharacters + ']{1}){1,' + previewWords + '}|.{0,' + previewCharacters + '})', 'i');
+      } else {
+        finder = new RegExp('(([^' + WORD_CHARACTERS + ']*[' + WORD_CHARACTERS + ']{1}){1,' + previewWords + '}|.{0,' + previewCharacters + '})', 'i');
         if (body) {
           textPreview = body.match(finder)[0] + '...';
         }
-      } else {
-        const wordMap = _.map(result.foundWords, function(count, word) {
-          return { word: word, count: count };
-        });
-        _.sortBy(wordMap, function(item) {
-          return item.count;
-        });
-        let wordIndex = 0;
-        let wordInPhraseStartPosition = lowerPhrase.indexOf(wordMap[wordIndex].word);
-        while (wordInPhraseStartPosition === -1) {
-          wordIndex++;
-          if (wordIndex === wordMap.length) throw new Error('search: cannot find word in phrase');
-          wordInPhraseStartPosition = lowerPhrase.indexOf(wordMap[wordIndex].word);
-        }
-        const regex = new RegExp('(([^' + wordCharacters + ']*[' + wordCharacters + ']{1}){1,' + previewWords + '}|.{0,' + previewCharacters + '})' + SearchAlgorithm._regularExpressions.escapeRegExp(wordMap[wordIndex].word) + '(([' + wordCharacters + ']{1}[^' + wordCharacters + ']*){1,' + previewWords + '}|.{0,' + previewCharacters + '})', 'i');
-        const snippet = phrase.match(regex)[0];
-        const snippetIndexInPhrase = phrase.indexOf(snippet);
-        if (snippet.length === phrase.length) {
-          textPreview = snippet;
-        } else if (snippetIndexInPhrase === 0) {
-          textPreview = snippet + '...';
-        } else if (snippetIndexInPhrase + snippet.length === phrase.length) {
-          textPreview = '...' + snippet;
-        } else {
-          textPreview = '...' + snippet + '...';
-        }
       }
 
-    } else {
-      finder = new RegExp('(([^' + wordCharacters + ']*[' + wordCharacters + ']{1}){1,' + previewWords + '}|.{0,' + previewCharacters + '})', 'i');
-      if (body) {
-        textPreview = body.match(finder)[0] + '...';
-      }
-    }
-
-    const searchTitleTagged = tag(result.foundWords, searchTitle);
-    const textPreviewTagged = tag(result.foundWords, textPreview);
-
-    return {
-      searchTitleTagged: searchTitleTagged,
-      searchTitle: searchTitle,
-      foundWords: foundWords,
-      textPreview: textPreview,
-      textPreviewTagged: textPreviewTagged,
-      id: result.model.get('_id')
-    };
-
-    function tag(words, text) {
-      let initial = '';
-      _.each(words, function(count, word) {
-        const wordPos = text.toLowerCase().indexOf(word);
-        if (wordPos < 0) return;
-        initial += text.slice(0, wordPos);
-        const highlighted = text.slice(wordPos, wordPos + word.length);
-        initial += "<span class='is-found'>" + highlighted + '</span>';
-        text = text.slice(wordPos + word.length, text.length);
-      });
-      initial += text;
-      return initial;
-    }
-  }
-
-  stripTags (text) {
-    text = $('<span>' + text + '</span>').html();
-    return text
-      .replace(replaceEndTagsRegEx, ' ')
-      .replace(replaceTagsRegEx, '')
-      .replace(replaceEscapedEndTagsRegEx, ' ')
-      .replace(replaceEscapedTagsRegEx, '');
+      const searchTitleTagged = tag(result.foundWords, searchTitle);
+      const textPreviewTagged = tag(result.foundWords, textPreview);
+      const id = result.model.get('_id');
+      return {
+        searchTitleTagged,
+        searchTitle,
+        foundWords,
+        textPreview,
+        textPreviewTagged,
+        id
+      };
+    });
+    searchObject.formattedResults = formattedResults;
+    return searchObject;
   }
 
   renderResults(results) {
@@ -186,7 +114,6 @@ export default class SearchResultsView extends Backbone.View {
   navigateToResultPage(event) {
     event && event.preventDefault();
     const blockID = $(event.currentTarget).attr('data-id');
-
     Adapt.navigateToElement('.' + blockID);
     Adapt.trigger('drawer:closeDrawer');
   }
