+++ less/search.less
@@ -17,10 +17,6 @@
     }
   }
 
-  &__item.drawer__item-btn {
-    padding-left: 2em;
-  }
-
   &__items-container {
     border-top: 1px solid @drawer-item-hover;
