+++ less/searchItem.less
@@ -1,8 +1,12 @@
 .search {
   &__item.no-results {
-    padding: @item-padding;
+    padding: @item-padding + 1em;
     line-height: @title-line-height;
     border-bottom: 1px solid @drawer-item-hover;
     color: @drawer-item-inverted;
+  }
+
+  &__item.drawer__item-btn {
+    padding-left: @item-padding + 1em;
   }
 }
